Moved format actions to render.tsx and added color to actios.
